+++ docker/docker-compose-localnet.yaml
@@ -23,7 +23,7 @@ services:
       - 5001:5001
 
   lotus:
-    image: textile/lotus-devnet:v0.4.0
+    image: textile/lotus-devnet:ntwk-calibration-7.29.0
     ports:
       - 7777:7777
     environment:
