+++ util/util.go
@@ -16,11 +16,17 @@ var (
 	AvgBlockTime = time.Second * 30
 )
 
-// CidUndef is a magic value to represent an undefined cid as a string.
-const CidUndef = "CID_UNDEF"
+const (
+	// MinDealDuration is the minium deal duration accepted in the Filecoin network.
+	// Original calculation: 180 * EpochsInADay
+	MinDealDuration = 180 * (24 * 60 * 60 / 25)
 
-// DefaultCidUndef is the string generated by the cid module's String() func for an undefined cid.
-const DefaultCidUndef = "b"
+	// CidUndef is a magic value to represent an undefined cid as a string.
+	CidUndef = "CID_UNDEF"
+
+	// DefaultCidUndef is the string generated by the cid module's String() func for an undefined cid.
+	DefaultCidUndef = "b"
+)
 
 // TCPAddrFromMultiAddr converts a multiaddress to a string representation of a tcp address.
 func TCPAddrFromMultiAddr(maddr ma.Multiaddr) (string, error) {
